Transmit gunangles as a short in rerelease protocol

+++ inc/common/msg.h
@@ -55,7 +55,7 @@ typedef struct {
     int16_t         viewangles[3];
     int16_t         viewoffset[3];
     int16_t         kick_angles[3];
-    int8_t          gunangles[3];
+    int16_t         gunangles[3];
     int8_t          gunoffset[3];
     uint16_t        gunindex;
     uint8_t         gunframe;
